requirements.txt
gspread
pandas
google-auth
google-auth-oauthlib
google-api-python-client
streamlit
streamlit-js-eval
gspread-dataframe
